keep PC matrix 2-dimensional even if only one PC
The prior code will fail when only one PC is selected because by default R will drop the singleton dimension and then the result is not an array type, so `dimnames(PCs)` generates an error. adding `drop = FALSE` to the subset keeps the singe PC as a column matrix, so everything will work.